Refactor LinkGR and implementations

+++ src/main/java/edu/city/studentuml/model/graphical/AbstractLinkGR.java
@@ -11,8 +11,8 @@ import java.util.Vector;
  */
 public abstract class AbstractLinkGR extends GraphicalElement {
     
-    public static final int REFLECTIVE_UP = 3;
-    public static final int REFLECTIVE_RIGHT = 6;
+    public static final double REFLECTIVE_UP = 3.0;
+    public static final double REFLECTIVE_RIGHT = 6.0;
 
     protected static Vector<AbstractLinkGR> linkInstances = new Vector<>();
 
@@ -147,14 +147,14 @@ public abstract class AbstractLinkGR extends GraphicalElement {
 
 
     public Point2D getCentreRoleA() {
-        double x = getTopLeftXA() + getWidthA() / 2;
-        double y = getTopLeftYA() + getHeightA() / 2;
+        double x = getTopLeftXA() + getWidthA() / 2.0;
+        double y = getTopLeftYA() + getHeightA() / 2.0;
         return new Point2D.Double(x, y);
     }
 
     public Point2D getCentreRoleB() {
-        double x = getTopLeftXB() + getWidthB() / 2;
-        double y = getTopLeftYB() + getHeightB() / 2;
+        double x = getTopLeftXB() + getWidthB() / 2.0;
+        double y = getTopLeftYB() + getHeightB() / 2.0;
         return new Point2D.Double(x, y);
     }
 
@@ -166,7 +166,7 @@ public abstract class AbstractLinkGR extends GraphicalElement {
         double angle;
 
         if (x2 - x1 != 0) {
-            double gradient = ((double) (y2 - y1)) / ((double) (x2 - x1));
+            double gradient = (y2 - y1) / (x2 - x1);
 
             if (x2 - x1 > 0) // positive gradient
             {
@@ -222,17 +222,17 @@ public abstract class AbstractLinkGR extends GraphicalElement {
         } else // reflective
         {
             int step = getReflectiveStep();
-            Rectangle2D definingRect1 = new Rectangle2D.Double(getXA() + step/2, getYB(), REFLECTIVE_RIGHT * step, step);
-            Rectangle2D definingRect2 = new Rectangle2D.Double(getXA() - step/2, getYA() - REFLECTIVE_UP * step, step, REFLECTIVE_UP * step);
-            Rectangle2D definingRect3 = new Rectangle2D.Double(getXA(), getYA() - REFLECTIVE_UP * step - step/2, REFLECTIVE_RIGHT * step, step);
-            Rectangle2D definingRect4 = new Rectangle2D.Double(getXA() + REFLECTIVE_RIGHT * step, getYA() - REFLECTIVE_UP * step - step/2, step, step * REFLECTIVE_UP + getYB() - getYA());
+            Rectangle2D definingRect1 = new Rectangle2D.Double(getXA() + step/2.0, getYB(), REFLECTIVE_RIGHT * step, step);
+            Rectangle2D definingRect2 = new Rectangle2D.Double(getXA() - step/2.0, getYA() - REFLECTIVE_UP * step, step, REFLECTIVE_UP * step);
+            Rectangle2D definingRect3 = new Rectangle2D.Double(getXA(), getYA() - REFLECTIVE_UP * step - step/2.0, REFLECTIVE_RIGHT * step, step);
+            Rectangle2D definingRect4 = new Rectangle2D.Double(getXA() + REFLECTIVE_RIGHT * step, getYA() - REFLECTIVE_UP * step - step/2.0, step, step * REFLECTIVE_UP + getYB() - getYA());
 
             return (definingRect1.contains(p) || definingRect2.contains(p) || definingRect3.contains(p) || definingRect4.contains(p));
         }
 
     }
 
-    // do not respond to drag-and-drop events;
+    // do not respond to drag-and-drop events
     // the links' positions depend on the classes that are connected
     public void move(int x, int y) {
         // empty
